+++ packages/expo-router/src/layouts/Tabs.tsx
@@ -1,9 +1,8 @@
-import { Screen } from '../views/Screen';
 import Tabs from './TabsClient';
 
-export * from '../react-navigation/bottom-tabs';
+export type { BottomTabNavigatorProps } from './TabsClient';
 
-Tabs.Screen = Screen;
+export * from '../react-navigation/bottom-tabs';
 
 export { Tabs };