+++ packages/expo-router/src/react-navigation/bottom-tabs/types.tsx
@@ -11,14 +11,12 @@ import type { EdgeInsets } from 'react-native-safe-area-context';
 
 import type { HeaderOptions, PlatformPressable } from '../elements';
 import type {
-  DefaultNavigatorOptions,
   Descriptor,
   NavigationProp,
   ParamListBase,
   RouteProp,
   TabActionHelpers,
   TabNavigationState,
-  TabRouterOptions,
   Theme,
 } from '../native';
 
@@ -458,14 +456,3 @@ export type BottomTabBarButtonProps = Omit<
   style?: StyleProp<ViewStyle>;
   onPress?: (e: React.MouseEvent<HTMLAnchorElement, MouseEvent> | GestureResponderEvent) => void;
 };
-
-export type BottomTabNavigatorProps = DefaultNavigatorOptions<
-  ParamListBase,
-  string | undefined,
-  TabNavigationState<ParamListBase>,
-  BottomTabNavigationOptions,
-  BottomTabNavigationEventMap,
-  BottomTabNavigationProp<ParamListBase>
-> &
-  TabRouterOptions &
-  BottomTabNavigationConfig;